Move some functions from dataflow into simulation-utilities.
