+++ src/plugins/dataflow/model/utilities/simulated-channel.ts
@@ -3,18 +3,11 @@
 // A sensor using a simulated channel uses the shared variable's value as its own on every tick.
 
 import { VariableType } from "@concord-consortium/diagram-view";
+
 import { NodeChannelInfo } from "./channel";
+import { inputVariableNamePart } from "../../../shared-variables/simulations/simulation-utilities";
 
 export const kSimulatedChannelType = "simulated-channel";
-export const kInputVariablePrefix = "input_";
-
-export function isInputVariable(variable: VariableType) {
-  return variable.name?.startsWith(kInputVariablePrefix);
-}
-
-function inputVariableNamePart(variable: VariableType) {
-  return variable.name?.slice(kInputVariablePrefix.length);
-}
 
 function simulatedChannelId(variable: VariableType) {
   return `SIM${inputVariableNamePart(variable)}`;
