+++ src/plugins/dataflow/model/utilities/simulated-output.ts
@@ -1,16 +1,8 @@
 import { Node } from "rete";
 import { VariableType } from "@concord-consortium/diagram-view";
+
 import { getOutputType } from "../../nodes/utilities/live-output-utilities";
-
-export const kOutputVariablePrefix = "output_";
-
-export function isOutputVariable(variable: VariableType) {
-  return variable.name?.startsWith(kOutputVariablePrefix);
-}
-
-function outputVariableNamePart(variable: VariableType) {
-  return variable.name?.slice(kOutputVariablePrefix.length);
-}
+import { outputVariableNamePart } from "../../../shared-variables/simulations/simulation-utilities";
 
 // Returns possible names to match with a shared variable based on the given node's outputType.
 // This is so output_LightBulb, output_lightbulb, output_light_bulb, etc will all match.
