+++ src/plugins/simulator/model/simulator-content.ts
@@ -2,12 +2,12 @@ import { reaction } from "mobx";
 import { types, Instance, getType, addDisposer, getSnapshot } from "mobx-state-tree";
 import { VariableSnapshot } from "@concord-consortium/diagram-view";
 
+import { kSimulatorTileType } from "../simulator-types";
+import { kSharedVariablesID, SharedVariables, SharedVariablesType } from "../../shared-variables/shared-variables";
 import { kBrainwavesKey } from "../simulations/brainwaves-gripper";
 import { simulations } from "../simulations/simulations";
 import { ITileExportOptions } from "../../../models/tiles/tile-content-info";
 import { TileContentModel } from "../../../models/tiles/tile-content";
-import { kSharedVariablesID, SharedVariables, SharedVariablesType } from "../../shared-variables/shared-variables";
-import { kSimulatorTileType } from "../simulator-types";
 import { SharedModelType } from "../../../models/shared/shared-model";
 
 export function defaultSimulatorContent(): SimulatorContentModelType {
